fix: move the callback example to example file
